Facilitate testing in isolation by adding interfaces for repositories.
+ Rename AlarmsDatabaseRepository -> RealAlarmsDatabaseRepository.
  Add AlarmsDatabaseRepository interface.
+ Rename HighlightsDatabaseRepository -> RealHighlightsDatabaseRepository.
  Add HighlightsDatabaseRepository interface.
+ Rename MetaDatabaseRepository -> RealMetaDatabaseRepository.
  Add MetaDatabaseRepository interface.
+ Rename SessionsDatabaseRepository -> RealSessionsDatabaseRepository.
  Add SessionsDatabaseRepository interface.
+ Rename EngelsystemNetworkRepository -> RealEngelsystemNetworkRepository.
  Add EngelsystemNetworkRepository interface.
+ Rename ScheduleNetworkRepository -> RealScheduleNetworkRepository.
  Add ScheduleNetworkRepository interface.
+ Rename SharedPreferencesRepository -> RealSharedPreferencesRepository.
  Add SharedPreferencesRepository interface.
